// packages/apps-vtex/src/utils/types.ts
export interface IProduct {
	productId: string;
	productName: string;
	brand: string;
	brandId: number;
	brandImageUrl?: string;
	cacheId?: string;
	linkText: string;
	productReference: string;
	categoryId: string;
	clusterHighlights: Record<string, unknown>;
	productClusters: Record<string, string>;
	categories: string[];
	categoriesIds: string[];
	link: string;
	description: string;
	skuSpecifications?: SkuSpecification[];
	priceRange: PriceRange;
	specificationGroups: SpecificationGroup[];
	properties: Array<{ name: string; values: string[] }>;
	selectedProperties: Array<{ key: string; value: string }>;
	releaseDate: string;
	origin?: string;
}

export type Product = IProduct & {
	items: Item[];
	clusterHighlights: Array<{ id: string; name: string }>;
	productClusters: Array<{ id: string; name: string }>;
};

export type LegacyProduct = IProduct & {
	metaTagDescription: string;
	productTitle: string;
	items: LegacyItem[];
	allSpecifications: string[];
	allSpecificationsGroups?: string[];
};
// ...
export type LegacyItem = Omit<Item, "variations" | "videos"> & {
	variations: string[];
	Videos: string[];
} & Record<string, string[]>;

export interface Item {
	itemId: string;
	name: string;
	nameComplete: string;
	complementName: string;
	ean: string;
	estimatedDateArrival: string;
	referenceId?: Array<{ Key: string; Value: string }>;
	measurementUnit: string;
	unitMultiplier: number;
	modalType: string | null;
	images: Image[];
	videos: string[];
	variations: Array<{
		name: string;
		values: string[];
	}>;
	sellers: Seller[];
	attachments: Array<{
		id: number;
		name: string;
		required: boolean;
		domainValues: string;
	}>;
	isKit: boolean;
	kitItems?: Array<{
		itemId: string;
		amount: number;
	}>;
}
// ...
export interface SkuSpecification {
	field: SKUSpecificationField;
	values: SKUSpecificationValue[];
}
export interface SKUSpecificationValue {
	name: string;
	id?: string;
	fieldId?: string;
	originalName?: string;
}

export interface SKUSpecificationField {
	name: string;
	originalName?: string;
	id?: string;
}

export interface Price {
	highPrice: number | null;
	lowPrice: number | null;
}

export interface PriceRange {
	sellingPrice: Price;
	listPrice: Price;
}

export interface SpecificationGroup {
	name: string;
	originalName: string;
	specifications: Array<{
		name: string;
		originalName: string;
		values: string[];
	}>;
}